feat: wire fixture converters into CLI as aimock convert subcommand
